+++ telegram/document.go
@@ -8,5 +8,5 @@ type Document struct {
 	Thumb    *PhotoSize `json:"thumb,omitempty"`     // Optional. Document thumbnail as defined by sender
 	FileName string     `json:"file_name,omitempty"` // Optional. Original filename as defined by sender
 	MimeType string     `json:"mime_type,omitempty"` // Optional. MIME type of the file as defined by sender
-	FileSize int64      `json:"file_size,omitempty"` // Optional. File size
+	FileSize int        `json:"file_size,omitempty"` // Optional. File size
 }
