+++ telegram/passport_file.go
@@ -5,6 +5,6 @@ PassportFile represents a file uploaded to Telegram Passport. Currently all Tele
 */
 type PassportFile struct {
 	FileID   string `json:"file_id"`   // Unique identifier for this file
-	FileSize int64  `json:"file_size"` // File size
-	FileDate int64  `json:"file_date"` // Unix time when the file was uploaded
+	FileSize int    `json:"file_size"` // File size
+	FileDate int    `json:"file_date"` // Unix time when the file was uploaded
 }
